+++ .gitignore
@@ -24,3 +24,5 @@ classes
 
 # Files from Forge MDK
 forge*changelog.txt
+
+.vscode
