+++ src/main/java/tfar/quickstack/util/MessageUtils.java
@@ -1,0 +1,9 @@
+package tfar.quickstack.util;
+
+import net.minecraft.ChatFormatting;
+
+public class MessageUtils {
+    public static String red(String message) {
+        return ChatFormatting.RED + message + ChatFormatting.RESET;
+    }
+}
