--- src/main/java/tfar/quickstack/util/Utils.java
@@ -1,9 +1,0 @@
-package tfar.quickstack.util;
-
-import net.minecraft.world.item.ItemStack;
-
-public class Utils {
-  public static boolean isFavorited(ItemStack stack){
-    return stack.hasTag() && stack.getTag().getBoolean("favorite");
-  }
-}
